requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1
djangorestframework==3.5.3
drf-nested-routers==0.11.1
future==0.16.0
gunicorn==19.6.0
oauthlib==2.0.1
rcssmin==1.0.6
requests==2.12.4
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
static3==0.7.0
TwitterSearch==1.0.1
